% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDeconvolution_simulationfunctions.r
\name{calculateBIC}
\alias{calculateBIC}
\title{calculate BIC from the output of DeClust}
\usage{
calculateBIC(exprM, deClustoutput)
}
\arguments{
\item{exprM}{the bulky expression data as input for deClust functions}

\item{deClustoutput}{the output of deClust functions}
}
\value{
BIC value
}
\description{
calculate BIC from the output of DeClust
}
\examples{
 data(exprM);
library("doParallel");
cl<-makeCluster(5,type="FORK",outfile="");
registerDoParallel(cl);
r<-deClustFromMarkerlognormal(exprM,3);
BIC<-calculateBIC(exprM,r)
}
