% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDeconvolution_simulationfunctions.r
\name{deClustFromMarkerlognormal}
\alias{deClustFromMarkerlognormal}
\title{Full version of DeClust}
\usage{
deClustFromMarkerlognormal(exprM, k, maxsubtypestep = 20, maxstep = 20,
  seed = 1)
}
\arguments{
\item{exprM}{a gene by sample expression matrix (in the original expression scale, not log-transformed, has to be non-negative)}

\item{k}{number of clusters
@param maxsubtypestep the maximum number of iterations in the inner lay of optimization for sample clustering}

\item{maxstep}{the maximum number of iterations in the outer lay of optimization for the fraction of each compartment}

\item{seed}{the random seed set for kmeans clustering}
}
\value{
A named list with three components: subtype, subtypeprofileM and subtypefra
}
\description{
This function is a full version of DeClust. It uses nonlinear optimization assuming the gene expression follows log-normal distribution. It also contains iteractive optimization procedures.
}
\examples{
 data(exprM);
library("doParallel");
cl<-makeCluster(5,type="FORK",outfile="");
registerDoParallel(cl);
r<-deClustFromMarkerlognormal(exprM,3);
}
