% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{exprM}
\alias{exprM}
\title{bulky gene expression for testing}
\format{A matrix with 200 rows(genes) and 102 columns(samples)}
\usage{
exprM
}
\description{
bulky gene expression for testing
}
\keyword{datasets}
