% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SI_geneset}
\alias{SI_geneset}
\title{Immune and stromal markers from R package estimate.}
\format{A data frame with 2 rows and 142 columns:}
\usage{
SI_geneset
}
\description{
Immune and stromal markers from R package estimate.
}
\keyword{datasets}
