% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{TCGAdeClustprofileM}
\alias{TCGAdeClustprofileM}
\title{precomputed expression profiles for 13 TCGA datasets by DeClust}
\format{A matrix of 10890 rows(genes) and 87 columns(different cancer subtypes,and tissue-specific immune and stromal compartments)}
\usage{
TCGAdeClustprofileM
}
\description{
DeClust was run for each of the 13 TCGA datasets separately (see methods of the paper for details). After that, the estimated profiles for cancer subtypes and immune/stromal compartments were assembled together into one single matrix for down-stream analysis. Only genes existed in more than half of the 13 datasets were kept. In order to make profiles more comparable between different datasets (e.g.,removing batch effects), quantile normalization was applied to the assemebled profiles.
}
\keyword{datasets}
