% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{TCGAdeClustsubtype}
\alias{TCGAdeClustsubtype}
\title{precomputed subtypes for 13 TCGA datasets by DeClust}
\format{A list of 13 items. Each item corresponds to one TCGA dataset, and is a vector of subtype annotation for samples in that TCGA dataset.}
\usage{
TCGAdeClustsubtype
}
\description{
precomputed subtypes for 13 TCGA datasets by DeClust
}
\keyword{datasets}
